farmprices/requirements-dev.txt
-r requirements.txt
pytest>=8.0.0,<9
pytest-flask>=1.3.0,<2
